Fix type casting in FNV-1 and FNV-1a hash update methods for byte processing

+++ include/hash23/fnv_1.h
@@ -37,7 +37,7 @@ namespace hash23 {
         constexpr void update(T const &data) {
             for (auto const &byte : data) {
                 hash_ *= prime();
-                hash_ ^= byte;
+                hash_ ^= static_cast<std::size_t>(byte);
             }
         }
